+++ push_swap.h
@@ -20,22 +20,22 @@
 
 typedef struct s_stack
 {
-	int				num;
-	int				idx;
+	long				num;
+	long				idx;
 	struct s_stack	*next;
 	struct s_stack	*prev;
 }	t_stack;
 
 void	error(void);
-int		assign_idx(char **results, int num);
-t_stack	*new_node(int num, char **results);
+long	assign_idx(char **results, long num);
+t_stack	*new_node(long num, char **results);
 void	append_node(t_stack **head, t_stack *node);
 t_stack	**get_stack(char **results);
-int		is_duplicated(t_stack **a);
-int		is_valid_input(t_stack **a);
+long	is_duplicated(t_stack **a);
+long	is_valid_input(t_stack **a);
 t_stack	*lst_last(t_stack *lst);
-int		is_sorted(t_stack **a);
-int		stack_size(t_stack **a);
+long	is_sorted(t_stack **a);
+long	stack_size(t_stack **a);
 
 void	sa(t_stack **a);
 void	sb(t_stack **b);
@@ -49,11 +49,14 @@ void	rra(t_stack **a);
 void	rrb(t_stack **b);
 void	rrr(t_stack **a, t_stack **b);
 
-int		min_num(t_stack **stack);
-int		find_pos(t_stack **a, int value);
+long	min_num(t_stack **stack);
+long	find_pos(t_stack **a, long value);
 void	sort_three(t_stack **a);
 void	small_sort(t_stack **a);
-int		get_max_pos(t_stack *stack);
+long	get_max_pos(t_stack *stack);
 void	butterfly_sort(t_stack **a, t_stack **b);
+
+// void	check_argv(char **results);
+// void	check_invalid_num(char **argv);
 
 #endif
